adding more tests

+++ src/main/resources/db.sql
@@ -4,13 +4,13 @@ DROP TABLE IF EXISTS slideshow_image;
 DROP TABLE IF EXISTS proof_of_play;
 
 CREATE TABLE images (
-    id           BIGSERIAL PRIMARY KEY,
+    id           BIGSERIAL    PRIMARY KEY,
     duration     INTEGER      NOT NULL,
     url          VARCHAR(255) NOT NULL UNIQUE
 );
 
 CREATE TABLE slideshows(
-    id           INTEGER PRIMARY KEY,
+    id           BIGSERIAL     PRIMARY KEY,
     created_at   TIMESTAMP     NOT NULL,
     name         VARCHAR(255)  NOT NULL
 );
@@ -25,7 +25,7 @@ CREATE TABLE slideshow_image (
 );
 
 CREATE TABLE proof_of_play(
-    id           INTEGER PRIMARY KEY,
+    id           BIGSERIAL     PRIMARY KEY,
     image_id     INTEGER       NOT NULL,
     slideshow_id INTEGER       NOT NULL,
     played_at    TIMESTAMP     NOT NULL  DEFAULT CURRENT_TIMESTAMP
